+++ Tracker/Trackers/Models/TrackersCell.swift
@@ -1,0 +1,21 @@
+//
+//  TrackersCell.swift
+//  Tracker
+//
+//  Created by Сергей Денисенко on 10.12.2023.
+//
+
+import UIKit
+
+class TrackersCell: UICollectionViewCell {
+    static let identifier = "TrackersCell"
+
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+    }
+
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
